+++ src/__tests__/hooks/medium.useEventOperations.spec.ts
@@ -6,7 +6,7 @@ import {
   setupMockHandlerDeletion,
   setupMockHandlerUpdating,
 } from '../../__mocks__/handlersUtils.ts';
-import { events } from '../../__mocks__/response/events.json' assert { type: 'json' };
+import { events } from '../../__mocks__/response/events.json';
 import { useEventOperations } from '../../hooks/useEventOperations.ts';
 import { server } from '../../setupTests.ts';
 import { Event, EventForm } from '../../types.ts';
