+++ src/provider/combinedContext.tsx
@@ -1,8 +1,13 @@
 import React, { createContext, useContext } from 'react';
 
+import { useEventForm } from '../feature/addOrEdit/@hooks/useEventForm';
+import {
+  EventOperations,
+  useEventOperations,
+} from '../feature/addOrEdit/@hooks/useEventOperations';
 import { CalendarView, useCalendarView } from '../feature/calendar/@hooks/useCalendarView';
 
-const CombinedContext = createContext<CalendarView | null>(null);
+const CombinedContext = createContext<(CalendarView & EventOperations) | null>(null);
 
 export const useCombinedContext = () => {
   const context = useContext(CombinedContext);
@@ -18,7 +23,9 @@ type Props = {
 
 export function CombinedContextProvider({ children }: Props) {
   const calendar = useCalendarView();
+  const { editingEvent, setEditingEvent } = useEventForm();
+  const eventOperations = useEventOperations(Boolean(editingEvent), () => setEditingEvent(null));
 
-  const value = { ...calendar };
+  const value = { ...calendar, eventOperations, editingEvent, setEditingEvent, ...eventOperations };
   return <CombinedContext.Provider value={value}>{children}</CombinedContext.Provider>;
 }
